requirements.txt
python-telegram-bot>=21.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiofiles==23.2.1
aiohttp>=3.9.3
anthropic>=0.17.0
openpyxl>=3.1.0 